move some stuff to payments, surgery in progress
