+++ geeksurvey/urls.py
@@ -26,11 +26,8 @@ urlpatterns = [
     path('help/', views.help, name='help'),
     path('accounts/', include('allauth.urls')),
     path('admin/', admin.site.urls),
+    path('payments/', include('payments.urls')),
     path('paypal/', include('paypal.standard.ipn.urls'), name='paypal-ipn'),
-    path('paypal/success/', views.paypal_success, name='paypal-return'),
-    path('paypal/failure/', views.paypal_failure, name='paypal-cancel'),
-    path('paypal/fund-account/', views.profile_fund, name='paypal_fund'),
-    path('paypal/claim-funds/', views.profile_claim, name='paypal_claim'),
     path('participate/', views.participate, name='participate'),
     path('participate/find/', views.part_discover, name='part_discover'),
     path('study/<uuid:study_id>/', views.study_landing_page, name='study_landing_page'),
