+++ payments/templates/payments/claim.html
@@ -1,0 +1,22 @@
+{% extends "base.html" %}
+
+{% load crispy_forms_tags %}
+
+{% block content %}
+
+<h1>Claim Your Earnings</h1>
+<p> AMOUNT: {{ profile.balance }} </p>
+<p> (you cannot claim amounts less than $5) </p>
+<hr>
+
+<form method="POST" enctype="multipart/form-data">
+{% csrf_token %}
+<fieldset class="form-group">
+    {{ form|crispy }}
+</fieldset>
+<div class="form-group">
+    <button class="btn btn-outline-info" type="submit">Claim</button>
+</div>
+
+{% endblock %}
+
