+++ payments/templates/payments/failure.html
@@ -1,0 +1,7 @@
+
+{% extends "base.html" %}
+{% block content %}
+
+<h1>PAYPAL FAILURE</h1>
+
+{% endblock %}
