+++ payments/templates/payments/success.html
@@ -1,0 +1,7 @@
+
+{% extends "base.html" %}
+{% block content %}
+
+<h1>PAYPAL SUCCESS</h1>
+
+{% endblock %}
