+++ payments/templates/payments/test.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+
+{% block content %}
+TEST BANANA
+{% endblock %}
+
